fix startup issues

+++ components/event-receiver/org.wso2.carbon.event.receiver.core/src/main/java/org/wso2/carbon/event/receiver/core/internal/CarbonEventReceiverService.java
@@ -437,7 +437,7 @@ public class CarbonEventReceiverService implements EventReceiverService {
                         String eventAdapterType = eventReceiver.getEventReceiverConfiguration().getFromAdapterConfiguration().getType();
                         if (eventAdapterType.equals(dependency)) {
                             EventReceiverConfigurationFile receiverConfigurationFile = getEventReceiverConfigurationFile(eventReceiver.getEventReceiverConfiguration().getEventReceiverName(), tenantId);
-                            if (receiverConfigurationFile != null) {
+                            if (receiverConfigurationFile != null && eventAdapterService != null) {
                                 fileList.add(receiverConfigurationFile);
                                 eventAdapterService.destroy(eventReceiver.getEventReceiverConfiguration().getFromAdapterConfiguration().getName());
                             }
